% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_const.R
\name{a_const}
\alias{a_const}
\title{a_const.R}
\usage{
a_const(j)
}
\arguments{
\item{j}{integer (desired level)}
}
\value{
a_const
}
\description{
\code{a_const} Utility internal function
}
\author{
Anestis Antoniadis <anestisa@gmail.com>

Yannig  Goude <yannig.goude@edf.fr>
}

