% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdaptCB.R
\name{AdaptCB}
\alias{AdaptCB}
\title{AdaptCB.R}
\usage{
AdaptCB(obs, sig = FALSE, sigma = NULL, alpha = 0.05,
  family = "DaubExPhase", bmin = 0.7, Mmax = 2.1)
}
\arguments{
\item{obs}{raw observed data}

\item{sig}{logical; if TRUE noise sd sigma is known; default FALSE}

\item{sigma}{if sig is TRUE the sd of noise. Otherwise NULL.}

\item{alpha}{1-alpha = confidence level}

\item{family}{the basic family that the wavelet comes from. The choices are \bold{DaubExPhase} for Daubechies' extremal phase wavelets, \bold{DaubLeAsymm} for Daubechies' ``least-asymmetric'' wavelets, and \bold{Coiflets} for Coifman's wavelets.}

\item{bmin}{the smallest degree of smoothness of the underlying function}

\item{Mmax}{the largest Lipschitz constant of the underlying function}
}
\value{
a list containing
\item{hfwidth}{the half width of the confidence band}
\item{sigmahat}{estimated or given noise standard deviation}
\item{jhat}{projection level of final estimate}
}
\description{
\code{AdaptCB} Function used to evaluate  confidence bands for the 
wavelet based components of an additive model.
}
\author{
Anestis Antoniadis <anestisa@gmail.com>

Yannig  Goude <yannig.goude@edf.fr>
}

