% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_const.R
\name{b_const}
\alias{b_const}
\title{b_const.R}
\usage{
b_const(j, v_phi)
}
\arguments{
\item{j}{integer (desired level)}

\item{v_phi}{variance of phi}
}
\value{
b_const
}
\description{
\code{b_const} Utility internal function
}
\author{
Anestis Antoniadis <anestisa@gmail.com>

Yannig  Goude <yannig.goude@edf.fr>
}

