% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbase.R
\name{bbase}
\alias{bbase}
\title{bbase}
\usage{
bbase(x, xl = min(x), xr = max(x), nseg = 10, deg = 3)
}
\arguments{
\item{x}{the x grid over which each B-spline basis components is evaluated}

\item{nseg}{number of internal equispaced knots. The default B‐spline bases has \code{nseg + deg} basis functions.}

\item{deg}{the degree of the B-splines. By default they are cubic (deg=3): they consist 
of segments that are polynomials of degree three. Linear (or other degree) B‐splines 
are obtained by changing the parameter deg.}
}
\value{
B the resulting B spline bases (a matrix of size n x (nseg + deg))
}
\description{
bbase
}
\author{
Paul Eilers and Brian Marx
}

