% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_const.R
\name{c_const}
\alias{c_const}
\title{c_const.R}
\usage{
c_const(j, sigma_phi, sigma)
}
\arguments{
\item{j}{integer (desired level)}

\item{sigma_phi}{sd of phi}

\item{sigma}{sd of noise}
}
\value{
c_const
}
\description{
\code{c_const} Utility internal function
}
\author{
Anestis Antoniadis <anestisa@gmail.com>

Yannig  Goude <yannig.goude@edf.fr>
}

