% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cPsi.R
\name{cPsi}
\alias{cPsi}
\title{cPsi.R}
\usage{
cPsi(family = "DaubExPhase", filternumber = 5, beta = 2.2)
}
\arguments{
\item{family}{the basic family that the wavelet comes from. The choices are \bold{DaubExPhase} for Daubechies' extremal phase wavelets, \bold{DaubLeAsymm} for Daubechies' ``least-asymmetric'' wavelets, and \bold{Coiflets} for Coifman's wavelets.}

\item{filternumber}{selects the desired filter, an integer that takes a value dependent upon the chosen wavelet family.
It selects the smoothness of wavelet that you want to use for estimation
The boundary handling is \code{bc="periodic"}, asumming therefore that the analyzed functions are  periodic on their interval of definition.}

\item{beta}{the highest smoothness index bound for the function to be estimated}
}
\value{
the constant c_psi
}
\description{
\code{Cpsi} Function used to evaluate the constant c_psi used in the bias part of the confidence band size
This is an internal function of package gamwave which allows to compute a confidence bands for the 
wavelet based components of amn additive model. Typically users will want to modify the defaults 
if model fitting deosn't seem appropriate.
}
\author{
Anestis Antoniadis <anestisa@gmail.com>

Yannig  Goude <yannig.goude@edf.fr>
}

