% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyad.R
\name{dyad}
\alias{dyad}
\title{dyad.R}
\usage{
dyad(j)
}
\arguments{
\item{j}{integer (desired level)}
}
\value{
all location indices at scale j
}
\description{
\code{dyad} Utility function  that returns location indices at level j
}
\author{
Anestis Antoniadis <anestisa@gmail.com>

Yannig  Goude <yannig.goude@edf.fr>
}

