% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitPsplineNormal.R
\name{fitPsplineNormal}
\alias{fitPsplineNormal}
\title{fitPsplineNormal}
\usage{
fitPsplineNormal(x, y, xnew = NULL, nseg = 10, pord = 2, lambda = 1,
  plot = T, se = 2)
}
\arguments{
\item{x}{abcissae of data}

\item{y}{response}

\item{xnew}{a new xgrid  over which we evaluate the fitted function (for display purposes). If xnew=NULL then xnew=sort(x).}

\item{nseg}{the number interior knots for the B-splines basis functions (all splines are of order 4 (cubic (deg=3)))}

\item{pord}{order of difference penalty (defaults to 2).}

\item{lambda}{smoothness parameter}

\item{plot}{logical parameter. if plot=T the data and fit are ploted.}

\item{se}{the width of standard error bands (computed and ploted only when se > 0)}

\item{range.x}{the min and max values of x over which the basis functions are evaluated}
}
\value{
a list containing

  \item{xnew}{the new xgrid  over which we evaluate the fitted function}
  \item{fHat}{a vector containing \eqn{\widehat{f}(x)}}
  \item{beta}{vector of estimated P-splines coefficients}
  \item{xnew}{the xnew grid for plotting: values are sorted}
  \item{munew}{the estimate evaluated on xnew}
  \item{lambda}{the penalty parameter used in the fit}
  \item{cv}{the cross-validation error associated to lambda}
  \item{effdim}{the effective dimension}
  \item{ed.resid}{the estimated degrees of freedom for the fit}
  \item{sigma}{the estimated dispersion associated to the fit}
}
\description{
Direct scatterplot semi-parametric regression smoothing with Eilers P-splines and a single smoothing paprameter
}
\examples{
# Obtain scatterplot data corresponding to environmental
# data from the R package `lattice'. Set up plotting 
# grid, knots and smoothing parameter:
require(lattice) ; attach(environmental) 
x <- radiation ; y <- ozone^(1/3)
range.x=range(x)
ng <- 200
xnew <- seq(range.x[1],range.x[2],length=ng) 
nseg <- 15
lambda<-11000
# Display the fit:
par(mfrow=c(1,2))
fHatsplines<-fitPsplineNormal(x,y,xnew=xnew,nseg=nseg,lambda=lambda,se=2,plot=T)
# Another plot to check the outputs
plot(x,y,xlim=range(xnew),bty="l",type="n",xlab="radiation",
  ylab="cuberoot of ozone",main="Direct fit with a given value of smooth parameter", 
 col.main="green4")
lines(fHatsplines$xnew,fHatsplines$munew,col='blue',lwd=2)
points(x,y,pch=19,col="grey",cex=0.5)
points(fHatsplines$x,fHatsplines$fHat,col="blue")
detach(environmental)

}
\author{
Paul Eilers and Brian Marx, 2007

Anestis Antoniadis <anestisa@gmail.com>

Yannig  Goude <yannig.goude@edf.fr>
}

