% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitPSplines.R
\name{fitPSplines}
\alias{fitPSplines}
\title{fitPSplines}
\usage{
fitPSplines(x, y, xnew = NULL, nseg = 10, pord = 2, seqLambda = NULL)
}
\arguments{
\item{x}{the abcissae}

\item{y}{the response}

\item{xnew}{a new xgrid  over which we evaluate the fitted function (for display purposes)}

\item{nseg}{the number interior knots for the B-splines basis functions (all splines are of order 4 (cubic (deg=3)))}

\item{pord}{an integer with values  1 or 2. Order of difference penalty (Defaults to 2).}

\item{seqLambda}{is a vector of smoothing parameters. if seqLambda is not provided  a default grid is used.}
}
\value{
a list containing

\item{fHat}{an array containing \eqn{\widehat{f}(x)}}
\item{betaHat}{an array containing the estimated P-splines regression coefficients}
\item{lambda}{the (CV optimised or not) penalised parameter}
\item{errCV}{the CV errors corresponding to the seqLambda}
\item{lseqLambda}{the log10(seqLambda)}
}
\description{
Direct scatterplot semi-parametric regression smoothing with Eilers P-splines
}
\examples{
# Obtain scatterplot data corresponding to environmental
# data from the R package `lattice'. Set up plotting 
# grid, knots and smoothing parameter:
require(lattice) ; attach(environmental) 
x <- radiation ; y <- ozone^(1/3)
range.x=range(x)
ng <- 200
xnew <- seq(range.x[1],range.x[2],length=ng) 
nseg <- 15
seqLambda<-10^seq(-1,5,by=0.2)
fit<-fitPSplines(x,y,xnew=xnew,nseg=nseg,seqLambda=seqLambda)
# Display the fit:
par(mfrow=c(1,2))
plot(x,y,xlim=range(xnew),bty="l",type="n",xlab="radiation",
ylab="cuberoot of ozone",main="fit with gcv",col.main="green4")
lines(xnew,fit$fHat,col='green',lwd=2)
points(x,y,pch=19,cex=0.5,col="grey")
plot(fit$lseqLambda, fit$errCV, xlab = 'log10(lambda)', ylab = 'CV',type="l")
points(fit$lseqLambda[which.min(fit$errCV)], fit$errCV[which.min(fit$errCV)],pch=19)
detach(environmental)

}
\author{
Anestis Antoniadis <anestisa@gmail.com>

Yannig  Goude <yannig.goude@edf.fr>
}

