% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fTrue.R
\name{fTrue}
\alias{fTrue}
\title{fTrue}
\usage{
fTrue(x)
}
\arguments{
\item{x}{input vector at which the function is evaluated}
}
\value{
a vector containing the evaluations of fTrue at x
}
\description{
Compute a noiseless signal with smooth and unsmooth parts
}
\examples{
n <- 300
x<-seq(0,1,length=n)
x<-sort(x)
y<-fTrue(x)
plot(x,y,col="dodgerblue",type="l",cex=0.3,
    main="A function with smooth and unsmooth parts",
    col.main="green4",xlab="x",ylab="f(x)")

}
\author{
Anestis Antoniadis <anestisa@gmail.com>

Yannig  Goude <yannig.goude@edf.fr>
}

