% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.add.data.R
\name{generate.add.data}
\alias{generate.add.data}
\title{generate.add.data}
\usage{
generate.add.data(n = 300, ds = 4, dw = 4, ps = 2, pw = 3,
  sigma = 1, seed = 1, active = c(1, 2, 4, 5, 6))
}
\arguments{
\item{n}{number of observations (length of data)}

\item{ds}{total number of smooth components}

\item{dw}{total number of  no-smooth components}

\item{ps}{number of effective smooth components}

\item{pw}{number of effective no-smooth components}

\item{sigma}{standard deviation of noise}

\item{seed}{the seed number}

\item{active}{list of of active variables}
}
\value{
a list with the following components
\item{d}{total number of  components}
\item{p}{total number of effective components}
\item{active}{the used set of active variables}
\item{not_active}{the complement set of active}
\item{A}{the design matrix of all d regressors}
\item{g}{the matrix whose columns contain the centered active additive functions} 
  evaluated on the corresponding entries of A
\item{ytrue}{the true additive signal (centered)}
\item{sigma}{the sd used for the simulations} 
\item{y}{the simulated observations}
}
\description{
Simulates an additive noisy function to be fitted with hybrid wavelets and splines
}
\examples{
n<-300
ds<-4
dw<-4
ps<-2
pw<-3
sigma<-1
seed<-1
active<-c(1,2,4,5,6)
out<-generate.add.data(n=n,ds=ds,dw=dw,ps=ps,pw=pw,sigma=sigma,seed=seed,active=active)

}
\author{
Anestis Antoniadis <anestisa@gmail.com>

Yannig  Goude <yannig.goude@edf.fr>
}

