% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid.fit.R
\name{hybrid.fit}
\alias{hybrid.fit}
\title{hybrid.fit}
\usage{
hybrid.fit(formula, data, numlev = floor(log2(nrow(data)) - 2), Hybrid = F,
  penalty = "SCAD", filternum = 5, family = "DaubLeAsymm",
  resolution = 16384, conv.thresh = 1e-10, offset = 0,
  max.iter = ncol(data) * 5, trace = F, nfolds = 10, method = "fREML",
  selection = NA, slope.fdr = 0.1, grpreg.penalty = "grLasso")
}
\arguments{
\item{formula}{a formula}

\item{data}{dataset}

\item{Hybrid}{True if boundaries are taken care of with hybrid wavelets}

\item{penalty}{the nonlinear penalisation method for model selection}

\item{conv.thresh}{convergence stopping value for relative difference in backfitting iterations}

\item{max.iter}{maximum number of backfitting iteration}

\item{trace}{tracing the computations}

\item{nfolds}{nfolds (default = 10) number of cross validation folds}

\item{method}{penalisation method for group selection}
}
\value{
an estimated hybrid model
\item{gsp}{the splines fitted part}
\item{gmw}{the wavelets fitted part}
\item{fitted}{sum of the splines and wavelets parts}
\item{formula}{model formula}
\item{selectvariable}{variables selected if selection is made}
}
\description{
Estimate an hybrid additive model with a spline part and a wavelet part
}
\examples{
 An example of smooth functions (to be compared with splines)
f1 <- function(x) return(3*sin(2*pi*x^3))
f2 <- function(x) return(15* x *exp(-x**2))
n <- 200
set.seed(1)
x1 <- runif(n)
x1 <- (x1-min(x1))/(max(x1)-min(x1))
x2 <- runif(n)
x2 <- (x2-min(x2))/(max(x2)-min(x2))
g1 <- f1(x1)-mean(f1(x1))
g2 <- f2(x2)-mean(f2(x2))
y <- g1+g2+ rnorm(n)*.5
A <- cbind(x1,x2)
numlev = floor(log2(n))-1
offset = -1 # fit with standard wavelets or with hybrid wavelets (uncomment the appropriate lines)
gmw <- GAMwv(y, A,numlev=(numlev-1), penalty="SCAD", offset=offset, trace =T)
gmw <- GAMwv(y, A,numlev=numlev, penalty="SCAD", trace  = T)# for comparing with mgcv's gam
gmw <- GAMwv.fit(y ~ w(x1) + w(x2), numlev=numlev, penalty="SCAD",Hybrid=TRUE, data=cbind(y,A), trace  = T)
gmw <- hybrid.fit(y ~ w(x1) + s(x2), penalty="SCAD", offset=offset, data=cbind(y,A))
gmw <- hybrid.fit(y ~ w(x1) + s(x2), numlev=numlev, penalty="SCAD", Hybrid=TRUE, data=cbind(y,A)) # for comparing with mgcv's gam

}

