% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret.formula.R
\name{interpret.formula}
\alias{interpret.formula}
\title{interpret.formula}
\usage{
interpret.formula(formula)
}
\arguments{
\item{formula}{a model formula}
}
\value{
Information about the model:
\item{pf} {a model formula for the parametric part}
\item{smooth.spec} {a list of descriptors for each spline element: smooth.spec}
\item{wavelet.spec} {a list of descriptors for each wavelet element: wavelet.spec}
\item{response} {the response covariate of the model}
}
\description{
interpret a partial linear additive model formula (with wavelets and/or splines) of the generic form:
y~x0+x1+x3*x4+s(x5)+s(x6,x7)+w(x8)+w(x9, by=fac) + ...
}
\examples{
f1 <- function(x) return(3*sin(2*pi*x^3)) 
f2 <- function(x) return(15* x *exp(-x**2))
n <- 200
set.seed(1)
x1 <- runif(n)
x1<-(x1-min(x1))/(max(x1)-min(x1))
x2 <- runif(n)
x2<-(x2-min(x2))/(max(x2)-min(x2))
g1<-f1(x1)-mean(f1(x1))
g2<-f2(x2)-mean(f2(x2))
y <- g1+g2+rnorm(n)*.5
A<-cbind(x1,x2)
numlev=floor(log2(n))-1
fac<-c(rep(1,n/2),rep(2,n/2))
data <- cbind(y, x1, x2, fac)
formula <- "y ~ w(x1) + s(x2, k=10)" 
interpret.formula(formula)
formula <- "y ~ w(x1, by=fac)" 
interpret.formula(formula)

}

