% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LevelTest.R
\name{LevelTest}
\alias{LevelTest}
\title{LevelTest.R}
\usage{
LevelTest(wc, theta, thr, n)
}
\arguments{
\item{wc}{vector of wavelet coefficients at that level.}

\item{theta}{boundary of the null hypothesis}

\item{thr}{threshold for the rejection region}
}
\value{
1 if the null is rejected / = 0 if not rejected
}
\description{
\code{LevelTest} Function to perform the test of wavelet coefficients at a single level
Internal function useful for computing  a confidence bands for the 
wavelet based components of amn additive model.
}
\author{
Anestis Antoniadis <anestisa@gmail.com>

Yannig  Goude <yannig.goude@edf.fr>
}

