% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeanThr.R
\name{MeanThr}
\alias{MeanThr}
\title{MeanThr.R}
\usage{
MeanThr(theta, thr)
}
\arguments{
\item{theta}{the mean}

\item{thr}{the threshold}
}
\value{
the mean threshold
}
\description{
\code{MeanThr} Function used to compute the mean of a thresholded half normal r.v.e
This is an internal function of package gamwave which allows to compute a confidence bands for the 
wavelet based components of amn additive model.
}
\author{
Anestis Antoniadis <anestisa@gmail.com>

Yannig  Goude <yannig.goude@edf.fr>
}

