% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewGAMwv.R
\name{NewGAMwv}
\alias{NewGAMwv}
\title{NewGAMwv}
\usage{
NewGAMwv(y, A, numlev = 6, local = FALSE, penalty = "SCAD",
  filternum = 5, family = "DaubLeAsymm", resolution = 16384,
  ngrid = 256, conv.thresh = 1e-10, offset = 0, max.iter = ncol(A) * 5,
  trace = FALSE, nfolds = 10)
}
\arguments{
\item{y}{the y observed values}

\item{A}{regressor matrix (without constants)}

\item{numlev}{number of levels for the wavelet decomposition}

\item{local}{logical parameter: TRUE if hybrid spline-coiflets are used ; FALSE if not (default = FALSE)}

\item{penalty}{type of penalty used for the penalised regression, could be "MCP" or "SCAD" or "lasso"}

\item{filternum}{selects the desired filter, an integer that takes a value dependent upon the chosen wavelet family.
It selects the smoothness of wavelet that you want to use in the decomposition}

\item{family}{the basic family that the wavelet comes from. The choices are \bold{DaubExPhase} for Daubechies' extremal phase wavelets, \bold{DaubLeAsymm} for Daubechies' ``least-asymmetric'' wavelets, and \bold{Coiflets} for Coifman's wavelets.}

\item{resolution}{the highest resolution wavelets in the expansion}

\item{ngrid}{the size of the xgrid over which to evaluate the fitted model}

\item{conv.thresh}{convergence stopping value for relative difference in backfitting iterations}

\item{offset}{integer offset for the estimated penalty parameter (fine tunning)}

\item{max.iter}{maximum number of backfitting iteration}
}
\value{
a gamwasp object containing

\item{alpha0}{a real number estimate of the mean}
\item{betaHat}{a vector of estimated regression parameters}
\item{Shat}{the estimated non-linear effects}
\item{type}{the type of basis functions used for the decomposition, "wavelet" or "splines"}
\item{Yhat}{the estimated non-linear effects evaluated on an equidistant design of size ngrid when \code{local=T}}
\item{Z}{the wavelet basis functions design matrix when \code{local=F}}
}
\description{
fit a wavelet based additive model with backfitting
}
\examples{
# An example of smooth functions (to be compared with splines)
f1 <- function(x) return(3*sin(2*pi*x^3))
f2 <- function(x) return(15* x *exp(-x**2))

n <- 200
set.seed(1)
x1 <- runif(n)
x1<-(x1-min(x1))/(max(x1)-min(x1))
x2 <- runif(n)
x2<-(x2-min(x2))/(max(x2)-min(x2))
g1<-f1(x1)-mean(f1(x1)) 
g2<-f2(x2)-mean(f2(x2))

y <- g1+g2+ rnorm(n)*.5 

A<-cbind(x1,x2)
numlev=floor(log2(n))-1
offset=-1
# fit with standard wavelets or with hybrid wavelets (uncomment the appropriate lines)
gmw<-NewGAMwv(y,A,numlev=numlev,penalty="SCAD",local=TRUE,offset=offset, ngrid=128)
# for comparing with mgcv's gam
gfit<-gam(y~s(x1)+s(x2)); plotdata<-plot(gfit,pages=1)
par(mfrow=c(1,2))
o<-order(x1)
plot(x1[o],y[o]-g2[o],type="n",xlab="True (dashed); gam (green), wavelet (purple)",ylab="y-g2(.)")
lines(x1[o],g1[o],lty='dashed',lwd=2)
points(x1[o],y[o]-g2[o],pch=19,cex=.5)
lines(x1[o],gmw$Shat[o,1],col='purple',lwd=2)
lines(plotdata[1][[1]]$x,plotdata[1][[1]]$fit,lty=2,lwd=2,col="green")
o<-order(x2)
plot(x2[o],y[o]-g1[o],type="n",xlab="True (dashed); gam (green), wavelet (purple)",ylab="y-g1(.)")
lines(x2[o],g2[o],lty='dashed',lwd=2)
points(x2[o],y[o]-g1[o],pch=19,cex=.5)
lines(x2[o],gmw$Shat[o,2],col='purple',lwd=2)
lines(plotdata[2][[1]]$x,plotdata[2][[1]]$fit,lty=2,lwd=2,col="green")

}
\author{
Anestis Antoniadis <anestisa@gmail.com>

Yannig  Goude <yannig.goude@edf.fr>
}

