% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PosefitWav.R
\name{PosefitWav}
\alias{PosefitWav}
\title{PosefitWaw}
\usage{
PosefitWav(x, y, xnew = NULL, range.x = NULL, numLevels = 6,
  filterNumber = 5, resolution = 16384, family = "DaubExPhase",
  gamma = 4)
}
\arguments{
\item{x}{the observed design values}

\item{y}{the y observed values}

\item{xnew}{a new xgrid  over which we evaluate the fitted function (for display purposes)}

\item{range.x}{the min and max values of x over which the basis functions are computed}

\item{numLevels}{number of levels for the wavelet decomposition}

\item{filterNumber}{selects the desired filter, an integer that takes a value dependent upon the chosen wavelet family. It selects the smoothness of wavelet that you want to use in the decomposition.}

\item{resolution}{the highest resolution wavelets in the expansion (defaults to 16384)}

\item{family}{the basic family that the wavelet comes from. The choices are \bold{DaubExPhase} for Daubechies' extremal phase wavelets, \bold{DaubLeAsymm} for Daubechies' ``least-asymmetric'' wavelets, and \bold{Coiflets} for Coifman's wavelets.}

\item{gamma}{penalty for gamma lasso}
}
\value{
a list containing

\item{fHat}{a  vector containing \eqn{\widehat{f}(x)}}
\item{beta}{a vctor containing the estimated wavelet regression coefficients}
}
\description{
fit a function by penalised wavelet regression
using a gamma Lasso penalisation (package gamlr)
}
\examples{
require(mgcv)
n <- 300 
set.seed(1) 
x <- sort(runif(n)) 
ng <- 1001
xnew <- seq(0,1,length=ng)
numLevels <- floor(log2(n))-2
range.x <- c(0,1)
f<-fTrue(x); sd=sqrt(var(f)); snr=4; sigma=sd/snr;
y <- f + rnorm(n) 
fhat<-PosefitWav(x,y,xnew=xnew,range.x=range.x,numLevels=numLevels,filterNumber=5, resolution=16384,family="DaubLeAsymm",gamma=4)$fHat

plot(x,y,xlim=range(xnew),ylim=range(fhat),bty="l",type="n",xlab="True(black);
wavelet fit (red); gam(mag)", ylab="y",main="Wavelet fit")
points(x,y,pch=19)
lines(x,f,col="black",lwd=2)
lines(xnew,fhat,col='red',lwd=2)
lines(x,gam(y~s(x))$fitted.values,col="magenta",lwd=2)


}
\author{
Anestis Antoniadis <anestisa@gmail.com>

Yannig  Goude <yannig.goude@edf.fr>
}

