% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predGAMwv.R
\name{predGAMwv}
\alias{predGAMwv}
\title{predGAMwv}
\usage{
predGAMwv(object, newdata, na.action = na.pass, weights = 1, ...)
}
\arguments{
\item{newdata}{a regressor one-dimensional vector}

\item{g}{a fitted GAMwv object with only one (wavelet) predictor}
}
\value{
a list containing

\item{predictor}{forecast on newdata}
}
\description{
Prediction from fitted Wavelet fit model
}
\examples{
n <- 300 
set.seed(1) 
x <- sort(runif(n)) 
f<-fTrue(x); sd=sqrt(var(f)); snr=4; sigma=sd/snr;#
y <- f + rnorm(n)
fit<-NewGAMwv.fit(y~w(x), numlev =6, offset=-1,filternum = 5, family = "DaubLeAsymm",
  penalty = "MCP", data=cbind(y,x))
res <- predGAMwv(fit)

}
\author{
Anestis Antoniadis <anestisa@gmail.com>

Yannig  Goude <yannig.goude@edf.fr>
}

