% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.gamBF.R, R/predict.gamwasp.R
\name{predict.gamwasp}
\alias{predict.gamwasp}
\title{predict.gamwasp}
\usage{
\method{predict}{gamwasp}(g, newdata)

\method{predict}{gamwasp}(g, newdata)
}
\arguments{
\item{g}{a fitted gamBF object}

\item{newdata}{a regressor matrix (without constants)}

\item{g}{a fitted gamBF object}

\item{newdata}{a regressor matrix (without constants)}
}
\value{
a list containing

\item{S.forecast}{forecast term by term}
\item{forecast}{forecast of the target}

a list containing

\item{S.forecast}{forecast term by term}
\item{forecast}{forecast of the target}
}
\description{
Prediction from fitted GAM (backfitting) model

Prediction from fitted GAM (mgcv's gam, or GAMsp or GAMwv) model
}
\examples{
y<-rnorm(10)


require(stacking)
f1 <- function(x) return(3*sin(2*pi*x^3))
f2 <- function(x) return(15* x *exp(-x**2))

n <- 200
set.seed(1)
x1 <- runif(n)
x1<-(x1-min(x1))/(max(x1)-min(x1))
x1<-sort(x1)
x2 <- runif(n)
x2<-(x2-min(x2))/(max(x2)-min(x2))
x2<-sort(x2)
g1<-f1(x1)-mean(f1(x1)) 
g2<-f2(x2)-mean(f2(x2))

y <- g1+g2+ rnorm(n)*.5 

A<-cbind(x1,x2)
numlev=floor(log2(n))-2
gamW<-GAMwv(y,A,numlev=numlev,penalty="SCAD",Hybrid=FALSE)
gamMGCV<-gam(y~s(x1)+s(x2))
gamMGCV$fitted=predict(gamMGCV)
x1new <- runif(n)
x1new<-(x1new-min(x1new))/(max(x1new)-min(x1new))
x1new<-sort(x1new)
x2new <- runif(n)
x2new<-(x2new-min(x2new))/(max(x2new)-min(x2new))
x2new<-sort(x2new)
g1new<-f1(x1new)-mean(f1(x1new)) 
g2new<-f2(x2new)-mean(f2(x2new))
ynew <- g1new+g2new+ rnorm(n)*.5 
Anew=cbind(x1new,x2new)
Data<-data.frame(x1=x1new,x2=x2new)
gamMGCV$forecast=predict(gamMGCV,newdata=Data)
mape(y,gamMGCV$fitted)
mape(ynew,gamMGCV$forecast)
gamW$forecast<-predict(gamW,newdata=Anew)$forecast
mape(y,gamW$fitted)
mape(ynew,gamW$forecast)

}
\author{
Anestis Antoniadis <anestisa@gmail.com>

Yannig  Goude <yannig.goude@edf.fr>

Anestis Antoniadis <anestisa@gmail.com>

Yannig  Goude <yannig.goude@edf.fr>
}

