% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.hybrid.R
\name{predict.hybrid}
\alias{predict.hybrid}
\title{predicthybrid}
\usage{
\method{predict}{hybrid}(hybrid, newdata)
}
\arguments{
\item{hybrid}{an hybrid model fitted object}

\item{data}{dataset on which we forecast}
}
\value{
a vector with the forecast
}
\description{
hybrid model forecasting
}

