% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleData.R
\name{scaleData}
\alias{scaleData}
\title{scaleData}
\usage{
scaleData(X)
}
\arguments{
\item{X}{the data matrix}
}
\value{
the scaled design matrix
}
\description{
sscale a data matrix so that each column ranges from 0 to 1
}

