% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateGAMsp.R
\name{simulateGAMsp}
\alias{simulateGAMsp}
\title{simulateGAMsp}
\usage{
simulateGAMsp(n, p = 7, a = -2.5, b = 2.5, sigma = 1.5)
}
\arguments{
\item{n}{the length of y}

\item{p}{the number of additive functions (without the constant)}

\item{a}{lower bound of the support of the additive components (the same for all)}

\item{b}{upper bound of the support of the additive components (the same for all)}

\item{sigma}{sd of noise, s/n ratio = sd}
}
\value{
a simulated object containing
\item{X}{the scaled design matrix}
\item{y}{the noisy signal}
\item{ftrue}{a matrix whose columns are the true additive components}
}
\description{
simulates an additive noisy function to be fitted with gam
}
\examples{
n<-200
p<-7
a<--2.5
b<-2.5
sigma<-1.5
out<-simulateGAMsp(n=n,p=p,a=a,b=b,sigma=sigma)
X<-out$X
G<-out$ftrue
y<-out$y
par(mfrow=c(3,3))
o<-order(X[,1])
plot(X[o,1],G[o,1],ylim=range((y-7)/3),type='l',col="red",lwd=2)
points(X[o,1],G[o,1]+sigma*rnorm(n),col="grey",pch=19)
o<-order(X[,2])
plot(X[o,2],G[o,2],ylim=range((y-7)/3),type='l',col="red",lwd=2)
points(X[o,2],G[o,2]+sigma*rnorm(n),col="grey",pch=19)
o<-order(X[,3])
plot(X[o,3],G[o,3],ylim=range((y-7)/1.5),type='l',col="red",lwd=2)
points(X[o,3],G[o,3]+sigma*rnorm(n),col="grey",pch=19)
o<-order(X[,4])
plot(X[o,4],G[o,4],ylim=range((y-5)/2),type='l',col="red",lwd=2)
points(X[o,4],G[o,4]+sigma*rnorm(n),col="grey",pch=19)
o<-order(X[,5])
plot(X[o,5],G[o,5],ylim=range((y-7)/3),type='l',col="red",lwd=2)
points(X[o,5],G[o,5]+sigma*rnorm(n),col="grey",pch=19)
o<-order(X[,6])
plot(X[o,6],G[o,6],type='l',ylim=range((y-7)/3),col="red",lwd=2)
points(X[o,6],G[o,6]+sigma*rnorm(n),col="grey",pch=19)
o<-order(X[,7])
plot(X[o,7],G[o,7],type='l',ylim=range((y-7)/3),col="red",lwd=2)
points(X[o,7],G[o,7]+sigma*rnorm(n),col="grey",pch=19)

}
\author{
Anestis Antoniadis <anestisa@gmail.com>

Yannig  Goude <yannig.goude@edf.fr>
}

