% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplineD.R
\name{SplineD}
\alias{SplineD}
\title{SplineD}
\usage{
SplineD(x, range.x, intKnots, drv = 0)
}
\arguments{
\item{x}{the x vector over which the basis functions are evaluated}

\item{range.x}{the min and max values of x over which the basis functions are evaluated
(x is allowed to take values outside the interior knots).}

\item{intKnots}{the interior knots for the B-splines basis functions (all splines are of order 4 (cubic))}

\item{drv}{an integer with values between 0 and 3. The derivative of the given order \code{drv} is evaluated at the x positions. Defaults to zero.}
}
\value{
a design matrix with each column containing the evaluation of a B-splines basis function on x in canonical form
}
\description{
Evaluates the design matrix for the O'Sullivan-type splines
}
\examples{

n <- 300
x<-seq(0,1,length=n)
range.x<-c(0,1)
numIntKnots <- 25
intKnots <- quantile(unique(x),seq(0,1,length=numIntKnots+2))[-c(1,numIntKnots+2)]
X <- SplineD(x,range.x=range.x,intKnots=intKnots)
plot(x,X[,ncol(X)],type="n",xlab="x",ylab="Basis functions", main="O-Spline basis functions for 25 equispaced knots")
for (i in (1:ncol(X))) lines(x,X[,i],col=i)

}
\author{
Anestis Antoniadis <anestisa@gmail.com>

Yannig  Goude <yannig.goude@edf.fr>
}
\references{
Wand, M.P. and Ormerod, J.T. (2008). On semiparametric regression with
O-Sullivan penalized splines. \emph{Australian and New Zealand Journal of Statistics}, \strong{50}, pages 179-198.
}

