% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplineP.R
\name{SplineP}
\alias{SplineP}
\title{SplineP}
\usage{
SplineP(x, range.x, nseg = 10, bdeg = 3)
}
\arguments{
\item{x}{the x grid over which each B-spline basis components is evaluated}

\item{nseg}{number of internal equispaced knots. The default B‐spline bases has \code{nseg + deg} basis functions. 
If it is too much for your data a smaller number for nseg will get nicer results.}

\item{bdeg}{the degree of the B-splines. By default they are cubic (bdeg=3): they consist 
of segments that are polynomials of degree three. Linear (or other degree) B‐splines 
are obtained by changing the parameter bdeg.}
}
\value{
B the resulting B spline bases (a matrix of size n x (nseg + deg))
}
\description{
SplineP
}
\examples{

## S1: Generate x on a linear grid of length 101
n<-101
x = seq(0, 1, length = n)
B <- SplineP(x)
matplot(x, B, type = 'l', lty = 1, lwd = 2,main="Individual cubic Bsplines")
#  Less B-splines
B = SplineP(x, nseg = 5)
matplot(x, B, type = 'l', lty = 1, lwd = 2,main="Individual cubic Bsplines")


}
\author{
Anestis Antoniadis <anestisa@gmail.com>

Yannig  Goude <yannig.goude@edf.fr>
}
\references{
Eilers, P. H. C. and Marx, B. D. (1996). Flexible smoothing with B-splines and penalties. \emph{Statistical Science},\strong{11}, pages 89–102.
}

