% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TauPsi.R
\name{TauPsi}
\alias{TauPsi}
\title{TauPsi.R}
\usage{
TauPsi(family = "DaubExPhase", filternumber = 5)
}
\arguments{
\item{family}{the basic family that the wavelet comes from. The choices are \bold{DaubExPhase} for Daubechies' extremal phase wavelets, \bold{DaubLeAsymm} for Daubechies' ``least-asymmetric'' wavelets, and \bold{Coiflets} for Coifman's wavelets.}

\item{filternumber}{selects the desired filter, an integer that takes a value dependent upon the chosen wavelet family.
It selects the smoothness of wavelet that you want to use for estimation
The boundary handling is \code{bc="periodic"}, asumming therefore that the analyzed functions are  periodic on their interval of definition.}
}
\value{
the constant tau_psi
}
\description{
\code{TauPsi} Function used to evaluate the constant tau_psi used in the bias part of the confidence band size
This is an internal function of package gamwave which allows to compute a confidence bands for the 
wavelet based components of amn additive model.
}
\author{
Anestis Antoniadis <anestisa@gmail.com>

Yannig  Goude <yannig.goude@edf.fr>
}

