% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpower.R
\name{tpower}
\alias{tpower}
\title{tpower}
\usage{
tpower(x, t, p)
}
\arguments{
\item{x}{the x value where the truncated power function is evaluated.}

\item{t}{the truncation parameter}

\item{p}{the power exponent}
}
\description{
tpower
}
\author{
Paul Eilers
}

