% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w.R
\name{w}
\alias{w}
\title{w.R}
\usage{
w(x, by = NA)
}
\arguments{
\item{x}{covariate}

\item{by}{covariate to multiple by x}
}
\value{
wavelet information
\item{alpha0}{a real number estimate of the mean}
\item{betaHat}{a vector of estimated regression parameters}
\item{Shat}{the estimated non-linear effects}
\item{type}{the type of basis functions used for the decomposition, "wavelet" or "splines"}
}
\description{
\code{w} Function used in definition of wavelet terms within gamwave model formula. 
The function does not evaluate a wavelet - it exists purely to help set up a model using wavelet.
}
\examples{
set.seed(0)
n<-200;sig2<-4
x1 <- runif(n, 0, 1)
fac<-c(rep(1,n/2),rep(2,n/2)) # create factor
w(x1) 
w(x1, by=fac)

}

