% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WavD.R
\name{WavD}
\alias{WavD}
\title{WavD}
\usage{
WavD(x, range.x = range(x), numLevels = 6, filterNumber = 5,
  family = "DaubExPhase", resolution = 16384)
}
\arguments{
\item{x}{the x vector over which the basis functions are computed}

\item{range.x}{the min and max values of x over which the basis functions are computed}

\item{numLevels}{number of levels for the wavelet decomposition}

\item{filterNumber}{selects the desired filter, an integer that takes a value dependent upon the chosen wavelet family.
It selects the smoothness of wavelet that you want to use in the decomposition
The boundary handling is \code{bc="periodic"}, asumming therefore that the analyzed functions are  periodic on their interval of definition.}

\item{family}{the basic family that the wavelet comes from. The choices are \bold{DaubExPhase} for Daubechies' extremal phase wavelets, \bold{DaubLeAsymm} for Daubechies' ``least-asymmetric'' wavelets, and \bold{Coiflets} for Coifman's wavelets.}

\item{resolution}{the highest resolution wavelets in the expansion}
}
\value{
a design matrix containing the wavelet basis functions for the given input vector
}
\description{
Computes the wavelet basis function design matrix
}
\examples{
# loading the lattice package for the plot
require("lattice")
n <- 300
x<-seq(0,1,length=300)
range.x<-c(0,1)
Z <- WavD(x,range.x=range.x,numLevels=4,family="DaubLeAsymm",filterNumber=5)
B<-cbind(rep(1,n),Z)
df <- data.frame(time = x,B)
form <- as.formula(paste(paste(names(df)[- 1],  collapse = ' + '),'x',  sep = '~'))
xyplot(form,data = df,type = 'l', grid=TRUE,outer =TRUE,ylab="DaubLeAsymm",ylim=c(-4,4),scales=list(tick.number=10))

}
\author{
Anestis Antoniadis <anestisa@gmail.com>

Yannig  Goude <yannig.goude@edf.fr>
}

