% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet.var.R
\name{wavelet.var}
\alias{wavelet.var}
\title{wavelet.var}
\usage{
wavelet.var(formula, data)
}
\arguments{
\item{formula}{model formula}

\item{data}{dataset with explanatory variables}
}
\value{
the new dataset
}
\description{
Creation of the dataset for a wavelet model
}
\examples{

f1 <- function(x) return(3*sin(2*pi*x^3)); f2 <- function(x) return(15* x *exp(-x**2)); n <- 200; set.seed(1); x1 <- runif(n)
x1<-(x1-min(x1))/(max(x1)-min(x1)); x2 <- runif(n); x2<-(x2-min(x2))/(max(x2)-min(x2)); g1<-f1(x1)-mean(f1(x1))
g2<-f2(x2)-mean(f2(x2)); y <- g1+g2+ rnorm(n)*.5;  A<-cbind(x1,x2); numlev=floor(log2(n))-1;
fac<-c(rep(1,n/2),rep(2,n/2))
data <- cbind(y, x1, fac)
formula <- "y ~ w(x1)" ; wavelet.var(formula, data)
formula <- "y ~ w(x1, by=fac)" ; wavelet.var(formula, data)

}

