% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZDaub.R
\name{ZDaub}
\alias{ZDaub}
\title{ZDaub}
\usage{
ZDaub(x, range.x = range(x), numLevels = 6, filterNumber = 5,
  resolution = 16384, family = "DaubExPhase")
}
\arguments{
\item{x}{the x value over which calculated the basis}

\item{range.x}{the min and max value over which calculated the basis}

\item{numLevels}{the level of the decomposition}

\item{filterNumber}{...}

\item{resolution}{...}

\item{family}{the family of the wavelet basis}
}
\value{
a matrix containing the wavelet basis
}
\description{
Compute the Daubechy wavelet basis
}
\examples{
y<-rnorm(10)


}
\author{
Anestis Antoniadis <anestisa@gmail.com>

Yannig  Goude <yannig.goude@edf.fr>
}

