% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZOSull.R
\name{ZOSull}
\alias{ZOSull}
\title{ZOSull}
\usage{
ZOSull(x, range.x, intKnots, drv = 0)
}
\arguments{
\item{x}{the x observed values}

\item{range.x}{the min and max value over which calculated the basis}

\item{intKnots}{number of internal knots}

\item{drv}{derivative order (default 0)}
}
\value{
a matrix containing the spline basis
}
\description{
Computes the O'Sullivan spline basis
}
\examples{
y<-rnorm(10)


}
\author{
Anestis Antoniadis <anestisa@gmail.com>

Yannig  Goude <yannig.goude@edf.fr>
}

